Arduino Programmed ESP8226 IoT Cat Monitoring Project
This is an esp8266 based cat monitoring project. Monitors food and water
bowl visits and litter box visit durations for my cat. Plots real time
data for visits as well as daily totals to a Thingspeak channel, as well
as daily totals and bowl alerts to the Twitter feed @MyCatStats.